// Musikschule/ClientApp/src/app/models/kurse.ts
export class Kurse {
    id: string | undefined | null;
    name: string | undefined | null;
    inhalt: string | undefined | null;
    start: Date | undefined | null;
    dauer: number | undefined | null;
    teilnehmerGesamt: number | undefined | null;
    teilnehmerBelegt: number | undefined | null;
}